fixed tests after moving __init__ imports
