+++ src/PythonFunctions/__init__.py
@@ -1,23 +1,5 @@
-# Update checker mainly.
-
-"""Init for all the modules
-Lets us rename, skip declartion of stuff and more."""
-
+# Module update checker, based off the github file
 import os
-
-
-# Credit: Dragmine149
-from . import Board, SaveModules, colours, run
-from .Check import Check
-from .CleanFolderData import Clean
-from .Convert import LocationConvert as LocConv
-from .Encryption import Encryption
-from .IsDigit import IsDigit
-from .Message import Message
-from .PrintTraceback import PrintTraceback
-from .Save import save
-from .TerminalDisplay import Display
-from .watermark import main as Watermark
 
 
 canReadGlobal = True
