+++ tests/__init__.py
@@ -7,3 +7,13 @@ sys.path.append(os.getcwd())
 
 # pylint: disable=wrong-import-position
 from src import PythonFunctions
+from src.PythonFunctions import Board, Check, SaveModules, colours, run
+from src.PythonFunctions import CleanFolderData
+from src.PythonFunctions.Convert import LocationConvert as LocConv
+from src.PythonFunctions.Encryption import Encryption
+from src.PythonFunctions.IsDigit import IsDigit
+from src.PythonFunctions.Message import Message
+from src.PythonFunctions.PrintTraceback import PrintTraceback
+from src.PythonFunctions.Save import save
+from src.PythonFunctions.TerminalDisplay import Display
+from src.PythonFunctions.watermark import main as Watermark
